+++ src/main/java/com/vajrapulse/vortex/results/SuccessEvent.java
@@ -4,26 +4,7 @@ package com.vajrapulse.vortex.results;
  * Represents a successful request event.
  * 
  * @param <T> the type of request element
+ * @param data the data associated with the successful event
  */
-public class SuccessEvent<T> {
-    private final T data;
-    
-    /**
-     * Creates a new SuccessEvent.
-     * 
-     * @param data the data associated with the successful event
-     */
-    public SuccessEvent(T data) {
-        this.data = data;
-    }
-    
-    /**
-     * Gets the data associated with this successful event.
-     * 
-     * @return the data
-     */
-    public T getData() {
-        return data;
-    }
+public record SuccessEvent<T>(T data) {
 }
-